Adjust bitmap conversion
The old wrapper was broken as it didn't initialize the target bitmap
properly, leading to a segfault. This patch makes it the same interface
as Freetype which fixes the previously mentioned issue and also allows
us to tweak the parameters of the target bitmap before conversion. This
is especially handy if we want to also convert the pitch flow.